Iteration 32 (#690)
* [FIXED]inline assignment in the template class
* [IDE][FIXED] debug source path
* [ELC][FIXED] return inside lock
* [ADDED] #154 : system'threading: ManualEvent, AutoResetEvent, CountDownEvent
* [SAMPLE][ADDED] threadpool
* ppc64le : fixing iteratorMethodTest
* [FIXED] declaring a lambda function with template based argument
* [API][ADDED] system'threading : Task
* [ADDED] system'io'Directory : static getFiles / getFiles<string>
* [FIXED] calling a method directly with nil argument when allowed
* unit test : calling a method with nil - strong - direct call
* [FIXED] single dispatch of a private method
* [Samples][ADDED] tasks
* [IDE][ADDED] it is possible to set the target type for a single file project
* [FIXED] closure can be type-casted in the compile-time

+++ elenasrc3/engine/elenamachine.cpp
@@ -58,7 +58,7 @@ uintptr_t ELENAMachine :: createPermString(SystemEnv* env, ustr_t s, uintptr_t c
 
    ObjectPage* header = (ObjectPage*)(nameAddr - elObjectOffset);
    header->vmtPtr = classPtr;
-   header->size = nameLen | struct_mask;
+   header->size = (int)nameLen | struct_mask;
 
    return nameAddr;
 }
